+++ src/core/server/systems/vehicle.ts
@@ -11,6 +11,7 @@ import { vehicleFuncs } from '../extensions/Vehicle';
 import { getPlayersByGridSpace } from '../utility/filters';
 import { getClosestEntity } from '../utility/vector';
 
+import '../views/garage';
 import '../views/dealership';
 import './fuel';
 
@@ -428,41 +429,3 @@ alt.onClient(VEHICLE_EVENTS.SET_ENGINE, VehicleFunctions.toggleEngine);
 alt.on('playerEnteringVehicle', VehicleFunctions.entering);
 alt.on('playerEnteredVehicle', VehicleFunctions.enter);
 alt.on('playerLeftVehicle', VehicleFunctions.leave);
-
-function handleSpawn(player: alt.Player, index: number) {
-    if (!player.data.vehicles) {
-        return;
-    }
-
-    if (index <= -1) {
-        return;
-    }
-
-    if (index >= player.data.vehicles.length) {
-        return;
-    }
-
-    const vehicleData = player.data.vehicles[index];
-    if (!vehicleData) {
-        return;
-    }
-
-    vehicleFuncs.new.spawn(player, vehicleData as Vehicle);
-}
-
-function handleDespawn(player: alt.Player) {
-    if (!player.data.vehicles) {
-        return;
-    }
-
-    if (player.lastVehicleID === null || player.lastVehicleID === undefined) {
-        return;
-    }
-
-    const vehicle = alt.Vehicle.all.find((veh) => veh && veh.id === player.lastVehicleID);
-    if (!vehicle) {
-        return;
-    }
-
-    vehicleFuncs.new.despawn(vehicle.id, player);
-}